+++ cli/src/commands/git/remote/add.rs
@@ -18,6 +18,7 @@ use jj_lib::repo::Repo as _;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
+use crate::commands::git::FetchTagsMode;
 use crate::git_util::absolute_git_url;
 use crate::ui::Ui;
 
@@ -31,6 +32,10 @@ pub struct GitRemoteAddArgs {
     /// Local path will be resolved to absolute form.
     #[arg(value_hint = clap::ValueHint::Url)]
     url: String,
+
+    /// Configure when to fetch tags
+    #[arg(long, value_enum, default_value_t = FetchTagsMode::Included)]
+    fetch_tags: FetchTagsMode,
 }
 
 pub fn cmd_git_remote_add(
@@ -44,7 +49,7 @@ pub fn cmd_git_remote_add(
         workspace_command.repo().store(),
         &args.remote,
         &url,
-        Default::default(),
+        args.fetch_tags.as_fetch_tags(),
     )?;
     Ok(())
 }